+++ GeecsBluesky/geecs_bluesky/devices/ca/__init__.py
@@ -11,6 +11,7 @@ Requires the ``ca`` extra (``aioca``): ``poetry install --extras ca``.
 """
 
 from geecs_bluesky.devices.ca.action_signals import CaActionSignalFactory
+from geecs_bluesky.devices.ca.confirm import CaConfirmSettable
 from geecs_bluesky.devices.ca.generic_detector import CaGenericDetector
 from geecs_bluesky.devices.ca.motor import CaMotor
 from geecs_bluesky.devices.ca.settable import CaSettable
@@ -22,6 +23,7 @@ from geecs_bluesky.devices.ca.triggerable import CaAcqTimestampReadable, CaTrigg
 __all__ = [
     "CaAcqTimestampReadable",
     "CaActionSignalFactory",
+    "CaConfirmSettable",
     "CaGenericDetector",
     "CaMotor",
     "CaSettable",